+++ src/uxds/components/index.ts
@@ -1,1 +1,15 @@
 export { ButtonPrimary, type ButtonPrimaryProps } from "./ButtonPrimary";
+export {
+  AppointmentSummaryPill,
+  AppointmentSummaryStack,
+  type AppointmentSummaryPillProps,
+  type AppointmentSummaryStackProps,
+} from "./AppointmentSummaryPill";
+export {
+  BookAppointmentProgress,
+  BOOK_APPOINTMENT_PROGRESS_LABELS,
+  buildBookProgressSteps,
+  type BookAppointmentProgressProps,
+  type BookProgressStep,
+  type BookProgressStepState,
+} from "./BookAppointmentProgress";